To support target node prioritization described in https://redmine.dataone.org/issues/2622, add a number of stub methods to ReplicationManager including:
getPendingRequestFactors() - returns a map of <nodeid, factor> entries that report on whether or not a node is currently over its request threshold.

getFailureFactors() - returns a map of <nodeid, factor> entries that reports the degree to which each node is succeeding/failing in replication requests.

getBandwidthFactors() - returns a map of <nodeid, factor> entries that report on the throughput ability of the node.

prioritizeNodes() - calculates the overall target node score based on the above factors, and accounts for preferred node requests. Returns the sorted listed of node identifiers.

TODO: These are just stubs.  Calls to the ReplicationDao interface are needed to calculate the metrics, and a ValueComparator class is needed for proper sorting.